Cluster Occurrence Data (large dataframe)

Cluster a large occurrence dataframe by date with the option to group each cluster using a specified radius distance. Each of these clusters of data represents a site and a pair of centered coordinates for each site is generated.

Usage

clusterise_sites_large_dataframe(
  dataframe,
  cluster_min_length,
  day_split_min_length = 10,
  group_radius = 40075000
)

Arguments

dataframe

A dataframe with occurrence data for the chosen taxon and location.

cluster_min_length

The minimum number of observations in each cluster.

day_split_min_length

By default the function filters out days with fewer than 10 observations. This value adjusts the minimum threshold.

group_radius

An optional value to have sites grouped. Group radius is measured in metres.

Value

The function returns a 'clusterised object', which is a list containing two elements: The first element is a list of data clusters. The second element is a dataframe that includes centred coordinates for each site, group number, and date.

Examples

# clusterise sites for the entire Santander province of Colombia
Colombia_Santander_dataframe <- subset(Colombia, stateProvince == "Santander")

clusterised_Santander <- clusterise_sites_large_dataframe(

 dataframe = Colombia_Santander_dataframe,
 cluster_min_length = 30

)

print(clusterised_Santander[[2]])
#>    longitude latitude site_group       date
#> 1  -73.85596 7.363071          1 2021-03-18
#> 2  -73.85572 7.354056          1 2021-03-20
#> 3  -73.85396 7.345819          1 2021-03-21
#> 4  -73.85837 7.349370          1 2021-03-22
#> 5  -73.86697 7.385701          1 2021-03-28
#> 6  -73.93555 7.385895          1 2021-05-30
#> 7  -73.86602 7.365386          1 2021-05-31
#> 8  -73.86660 7.385468          1 2021-06-02
#> 9  -73.86355 7.371646          1 2021-06-03
#> 10 -74.20774 6.073438          1 2021-06-05
#> 11 -73.97160 6.915441          1 2021-06-06
#> 12 -73.87424 7.315279          1 2021-06-07
#> 13 -73.86336 7.253735          1 2021-07-01
#> 14 -73.87146 7.282708          1 2021-07-02
#> 15 -73.85599 7.249371          1 2021-07-03
#> 16 -73.87120 7.283191          1 2021-07-04
#> 17 -73.89066 7.310576          1 2021-07-05
#> 18 -73.77900 7.377343          1 2021-07-06
#> 19 -73.87974 7.233874          1 2021-07-07
#> 20 -73.80636 7.357476          1 2021-07-08
#> 21 -73.81867 7.395084          1 2021-07-09
#> 22 -73.86611 7.356642          1 2021-07-12
#> 23 -73.80135 7.321644          1 2021-07-13
#> 24 -73.85234 7.386549          1 2021-07-14
#> 25 -73.73319 7.321563          1 2021-07-15
#> 26 -73.86561 7.397551          1 2021-07-16
#> 27 -73.70409 7.310465          1 2021-07-17
#> 28 -73.86438 7.242329          1 2021-08-14
#> 29 -73.88775 7.249436          1 2021-08-15
#> 30 -73.83731 7.301789          1 2021-08-16
#> 31 -73.85183 7.292450          1 2021-08-18
#> 32 -73.84857 7.289851          1 2021-08-19
#> 33 -73.86154 7.244816          1 2021-08-20
#> 34 -73.89353 7.204895          1 2021-08-21
#> 35 -73.89336 7.207342          1 2021-08-22
#> 36 -73.90824 7.260450          1 2021-08-23
#> 37 -73.92073 7.264681          1 2021-08-25
#> 38 -73.89955 7.295290          1 2021-08-26
#> 39 -73.88663 7.313230          1 2021-08-27
#> 40 -73.89316 7.205097          1 2021-10-31
#> 41 -73.86561 7.263230          1 2021-11-01
#> 42 -73.87716 7.285349          1 2021-11-02
#> 43 -73.88662 7.270188          1 2021-11-03
#> 44 -73.87796 7.263157          1 2021-11-04
#> 45 -73.90940 7.260669          1 2021-11-05
#> 46 -73.88655 7.251194          1 2021-11-06
#> 47 -73.83450 7.278388          1 2021-11-08
#> 48 -73.87982 7.153337          1 2021-11-10
#> 49 -73.88027 7.093077          1 2021-11-13
#> 50 -73.85201 7.294357          1 2022-01-28
#> 51 -73.83671 7.273990          1 2022-01-29
#> 52 -73.83460 7.277904          1 2022-01-30
#> 53 -73.84257 7.287955          1 2022-01-31
#> 54 -73.84169 7.242024          1 2022-02-01
#> 55 -73.84487 7.256973          1 2022-02-02
#> 56 -73.84268 7.182902          1 2022-02-04
#> 57 -73.87554 7.230265          1 2022-02-05
#> 58 -73.88122 7.279388          1 2022-02-06
#> 59 -73.91400 7.271642          1 2022-02-08
#> 60 -73.84898 7.288068          1 2022-02-13
#> 61 -73.83055 7.274892          1 2022-02-14
#> 62 -73.82880 7.247969          1 2022-02-15
#> 63 -73.89167 7.205998          1 2022-02-16
#> 64 -73.85132 7.244891          1 2022-02-17
#> 65 -73.91807 7.288713          1 2022-02-18
#> 66 -73.85034 7.258010          1 2022-02-19
#> 67 -73.88076 7.265456          1 2022-02-20
#> 68 -73.89232 7.205226          1 2022-02-21
#> 69 -73.81809 7.295247          1 2022-02-27
#> 70 -73.86369 7.260309          1 2022-02-28
#> 71 -73.85891 7.352441          1 2022-03-01
#> 72 -73.89766 7.297196          1 2022-03-02
#> 73 -73.87665 7.314925          1 2022-03-03
#> 74 -73.90161 7.248095          1 2022-03-04
#> 75 -73.88275 7.259575          1 2022-03-05
#> 76 -73.85844 7.325502          1 2022-03-06
#> 77 -73.90313 7.119544          1 2022-03-07
#> 78 -73.83806 7.128784          1 2022-03-09
#> 79 -73.86330 7.248224          1 2022-03-17
#> 80 -73.86007 7.253641          1 2022-03-19
#> 81 -73.87185 7.288971          1 2022-03-21
#> 82 -73.87557 7.294555          1 2022-03-22
#> 83 -73.80460 7.363378          1 2022-03-23
#> 84 -73.84772 7.350171          1 2022-03-24
#> 85 -73.77536 7.373977          1 2022-03-25
#> 86 -73.83793 7.357674          1 2022-03-26
#> 87 -73.86730 7.276636          1 2022-03-28
#> 88 -73.85486 7.419271          1 2022-03-29
#> 89 -73.90716 7.272219          1 2022-03-30
#> 90 -73.84040 7.369485          1 2022-03-31
#> 91 -73.86821 7.395321          1 2022-04-01
#> 92 -73.82533 7.326988          1 2022-04-02
#> 93 -73.70768 7.314034          1 2022-04-04
#> 94 -73.70536 7.311770          1 2022-04-06