Filter Group by Size

Filter a clusterised object by removing groups below a specified size.

Usage

filter_groups_by_minsize(clusterised_object, min_group_size)

Arguments

clusterised_object

An object created by the clusterise_sites functions.

min_group_size

The minimum size of each group.

Value

A new clusterised object with groups of at least the specified minimum size.

Examples

# return groups that have at least 10 sites

Colombia_Huila_dataframe <- subset(Colombia, stateProvince == "Huila")

clusterised_Huila <- clusterise_sites(dataframe = Colombia_Huila_dataframe,
                                      cluster_min_length = 30,
                                      group_radius = 20000
)

clusterised_Huila_largegroups <-
  filter_groups_by_minsize(

    clusterised_object = clusterised_Huila,
    min_group_size = 10

)

print(clusterised_Huila_largegroups[[2]])
#>    longitude latitude site_group       date
#> 1  -76.13123 1.831492          4 2008-08-23
#> 2  -76.13957 1.827111          4 2008-09-15
#> 3  -76.14045 1.827529          4 2008-09-17
#> 4  -76.13981 1.827387          4 2008-09-19
#> 5  -76.13827 1.827043          4 2008-09-21
#> 6  -76.13812 1.827224          4 2008-09-23
#> 7  -76.13867 1.827080          4 2008-09-27
#> 8  -76.13942 1.827976          4 2008-09-29
#> 9  -76.13839 1.826807          4 2008-10-01
#> 10 -76.13344 1.829129          4 2008-10-03
#> 11 -76.13898 1.827263          4 2008-10-05
#> 12 -76.13924 1.827158          4 2008-10-07
#> 13 -76.13966 1.827432          4 2008-10-09
#> 14 -76.13952 1.827509          4 2008-10-11
#> 15 -76.13885 1.827348          4 2008-10-13
#> 16 -76.13998 1.826505          4 2008-10-15
#> 17 -76.13921 1.826915          4 2008-10-17
#> 18 -76.13944 1.827849          4 2008-10-19
#> 19 -76.13920 1.827229          4 2008-10-21
#> 20 -76.13993 1.827921          4 2008-10-23
#> 21 -76.13847 1.826702          4 2008-10-25
#> 22 -76.13928 1.827402          4 2008-10-27
#> 23 -76.14349 1.827900          4 2008-10-29
#> 24 -76.13890 1.827041          4 2008-10-31