Filter Group by Name

Filter a clusterised object that has been assigned groups and remove groups that are not a specified label.

Usage

filter_groups_by_number(clusterised_object, group_number)

Arguments

clusterised_object

An object created by the clusterise_sites function.

group_number

A number that identifies the group to be extracted.

Value

A new clusterised object that only contains sites from the specified group.

Examples

# extract groups 3, 5, and 7 from a clusterised object
Colombia_Caquetá_dataframe <- subset(Colombia, stateProvince == "Caquetá")

clusterised_Caquetá <- clusterise_sites(dataframe = Colombia_Caquetá_dataframe,
                                       cluster_min_length = 30,
                                       group_radius = 20000
)

print(clusterised_Caquetá[[2]])
#>    longitude latitude site_group       date
#> 1  -75.54658 1.558959          1 2008-06-01
#> 2  -75.81619 1.339552          2 2008-06-01
#> 3  -75.78822 1.289720          2 2016-03-08
#> 4  -75.78894 1.267924          2 2019-02-08
#> 5  -75.65327 1.441539          3 2016-03-09
#> 6  -75.66336 1.456000          3 2016-03-16
#> 7  -75.65286 1.441330          3 2019-01-24
#> 8  -75.65567 1.439890          3 2019-01-25
#> 9  -75.65288 1.385920          3 2019-02-06
#> 10 -75.97152 1.194183          4 2016-03-15
#> 11 -75.61600 1.709897          5 2016-03-22
#> 12 -75.61369 1.716690          5 2016-03-29
#> 13 -75.63117 1.678451          5 2016-03-30
#> 14 -75.63488 1.651888          5 2016-04-13
#> 15 -75.61261 1.717140          5 2018-12-17
#> 16 -75.65950 1.744000          5 2019-01-21
#> 17 -75.59994 1.707670          5 2019-01-23
#> 18 -75.63404 1.657868          5 2019-02-18
#> 19 -75.77772 1.438032          7 2016-05-13
#> 20 -75.76399 1.431014          7 2016-05-19
#> 21 -75.75257 1.403800          7 2018-11-07
#> 22 -75.86772 1.475750          7 2018-11-27
#> 23 -75.74022 1.509550          7 2018-11-28
#> 24 -75.80704 1.469540          7 2018-12-12
#> 25 -75.81220 1.510290          7 2018-12-14
#> 26 -75.88993 1.490180          7 2019-01-23
#> 27 -75.90423 1.688207          8 2017-01-25
#> 28 -75.90526 1.629135          8 2017-01-28
#> 29 -75.89100 1.614610          8 2017-01-30
#> 30 -75.01101 1.317384         12 2018-12-04
#> 31 -75.00592 1.321170         12 2019-02-14

Colombia_Caquetá_3 <- filter_groups_by_number(clusterised_Caquetá, 3)
Colombia_Caquetá_5 <- filter_groups_by_number(clusterised_Caquetá, 5)
Colombia_Caquetá_7 <- filter_groups_by_number(clusterised_Caquetá, 7)

print(Colombia_Caquetá_3[[2]])
#>   longitude latitude site_group       date
#> 5 -75.65327 1.441539          3 2016-03-09
#> 6 -75.66336 1.456000          3 2016-03-16
#> 7 -75.65286 1.441330          3 2019-01-24
#> 8 -75.65567 1.439890          3 2019-01-25
#> 9 -75.65288 1.385920          3 2019-02-06
print(Colombia_Caquetá_5[[2]])
#>    longitude latitude site_group       date
#> 11 -75.61600 1.709897          5 2016-03-22
#> 12 -75.61369 1.716690          5 2016-03-29
#> 13 -75.63117 1.678451          5 2016-03-30
#> 14 -75.63488 1.651888          5 2016-04-13
#> 15 -75.61261 1.717140          5 2018-12-17
#> 16 -75.65950 1.744000          5 2019-01-21
#> 17 -75.59994 1.707670          5 2019-01-23
#> 18 -75.63404 1.657868          5 2019-02-18
print(Colombia_Caquetá_7[[2]])
#>    longitude latitude site_group       date
#> 19 -75.77772 1.438032          7 2016-05-13
#> 20 -75.76399 1.431014          7 2016-05-19
#> 21 -75.75257 1.403800          7 2018-11-07
#> 22 -75.86772 1.475750          7 2018-11-27
#> 23 -75.74022 1.509550          7 2018-11-28
#> 24 -75.80704 1.469540          7 2018-12-12
#> 25 -75.81220 1.510290          7 2018-12-14
#> 26 -75.88993 1.490180          7 2019-01-23