Generate Diversity Index Statistics

Calculates diversity index values for each site in a clusterised object.

Usage

generate_stats(clusterised_object)

Arguments

clusterised_object

An object created by the clusterise_sites function.

Value

A dataframe where each row pertains to a site and each column has higher taxa labels, time & location data, as well as diversity index values.

Examples

# generate species count tables with a minimum of 30 individuals from 5 species
Colombia_Caquetá_dataframe <- subset(Colombia, stateProvince == "Caquetá")

clusterised_Caquetá <- clusterise_sites(dataframe = Colombia_Caquetá_dataframe,
                                       cluster_min_length = 30,
                                       group_radius = 20000
)

stats_Caquetá <- generate_stats(clusterised_Caquetá)

head(stats_Caquetá)
#>   longitude latitude site_group       date year month day  country
#> 1 -75.54658 1.558959          1 2008-06-01 2008     6   1 Colombia
#> 2 -75.81619 1.339552          2 2008-06-01 2008     6   1 Colombia
#> 3 -75.78822 1.289720          2 2016-03-08 2016     3   8 Colombia
#> 4 -75.78894 1.267924          2 2019-02-08 2019     2   8 Colombia
#> 5 -75.65327 1.441539          3 2016-03-09 2016     3   9 Colombia
#> 6 -75.66336 1.456000          3 2016-03-16 2016     3  16 Colombia
#>   stateProvince  kingdom     phylum   class   N  S        H         D      Dinv
#> 1       Caquetá Animalia Arthropoda Insecta 129 53 3.672516 0.9648459 28.446154
#> 2       Caquetá Animalia Arthropoda Insecta  81 37 3.414379 0.9605243 25.332046
#> 3       Caquetá Animalia Arthropoda Insecta  36  9 1.901906 0.8117284  5.311475
#> 4       Caquetá Animalia Arthropoda Insecta  49 24 2.970650 0.9396085 16.558621
#> 5       Caquetá Animalia Arthropoda Insecta  52 27 3.012315 0.9378698 16.095238
#> 6       Caquetá Animalia Arthropoda Insecta  34 19 2.787044 0.9256055 13.441860
#>           d         J day_length
#> 1 10.700001 0.9249990   12.20434
#> 2  8.192153 0.9455709   12.19245
#> 3  2.232443 0.8655948   12.09705
#> 4  5.909831 0.9347388   12.06910
#> 5  6.580207 0.9139758   12.09661
#> 6  5.104413 0.9465451   12.10555