Add to Individual Occurrence Map

Add range polygons to an existing interactive map.

Usage

mapAdd_rangePoly(
  existing_map,
  taxon_dataframe,
  site_name = "",
  colour = "#60f",
  site_legend = T,
  poly_opacity = 0.4
)

Arguments

existing_map

A map object that has been created using the map_start function.

site_name

An optional label to add the the coordinate pins.

colour

The colour of the coordinate pins.

site_legend

A boolean value to choose to add a site name legend to the map.

poly_opacity

An optional value to set the opacity of the polygon.

clusterised_object

An object created by the clusterise_sites functions.

Value

Coordinate pins to be added to an existing interactive map object.

Examples

# add site coordinates to an existing map
species_dataframe <- subset(Andes, species == "Boana crepitans")

map <- mapStart_rangePoly(species_dataframe, site_name = "Whole Dataset")

species_dataframe_split <- split(species_dataframe, species_dataframe$year)

map <- mapAdd_rangePoly(map, species_dataframe_split$`2011`, "2016", "purple")
map <- mapAdd_rangePoly(map, species_dataframe_split$`2013`, "2013", "green")
map <- mapAdd_rangePoly(map, species_dataframe_split$`2015`, "2015", "red")
map <- mapAdd_rangePoly(map, species_dataframe_split$`2017`, "2017", "blue")

map